Update requirements
Since version 1.7, onnx-tensorflow is exporting TF2.X SavedModels
that are not compatible with essentia-tensorflow yet.
https://github.com/onnx/onnx-tensorflow/releases/tag/v1.7.0

As the main goal of this repository is to provide reliable conversion
tools for the essentia tensorflow models
https://essentia.upf.edu/machine_learning.html#using-pre-trained-tensorflow-models
let's rely on the latest version exporting TF1.X .pb graphs (onnx-tensorflow==1.6)
for now.

TF is fixed to 2.2.1 as recommended from onnx-tensorflow.